+++ packages/slate-plugins/src/__test-utils__/types.ts
@@ -1,9 +1,0 @@
-import { Editor } from 'slate';
-
-export interface FixtureEditor {
-  module: {
-    run: Function;
-    input: Editor;
-    output: Editor;
-  };
-}