+++ packages/slate-plugins/src/__tests__/common/transforms/setPropsToNodes/descendant-text.spec.tsx
@@ -1,15 +1,16 @@
 /** @jsx jsx */
 
 import { jsx } from '__test-utils__/jsx';
-import { setPropsToDescendants } from 'common/transforms';
+import { isDescendant } from 'common/queries';
+import { setPropsToNodes } from 'common/transforms';
 
-const node = (<text>test</text>) as any;
+const node = (<htext>test</htext>) as any;
 
 const props = { a: 1 };
 
 const output = (<htext a={1}>test</htext>) as any;
 
 it('should set props to the text node', () => {
-  setPropsToDescendants(node, props);
+  setPropsToNodes(node, props, { filter: isDescendant });
   expect(node).toEqual(output);
 });