+++ packages/slate-plugins/src/__tests__/common/transforms/setPropsToNodes/elements-nested.spec.tsx
@@ -1,15 +1,16 @@
 /** @jsx jsx */
 
 import { jsx } from '__test-utils__/jsx';
-import { setPropsToElements } from 'common/transforms';
+import { setPropsToNodes } from 'common/transforms';
 import { ListType } from 'elements/list';
 import { PARAGRAPH } from 'elements/paragraph';
+import { Element } from 'slate';
 
 const node = (
-  <li>
+  <hli>
     test
-    <p>test</p>test
-  </li>
+    <hp>test</hp>test
+  </hli>
 ) as any;
 
 const props = { a: 1 };
@@ -25,6 +26,6 @@ const output = (
 ) as any;
 
 it('should set props to all elements', () => {
-  setPropsToElements(node, props);
+  setPropsToNodes(node, props, { filter: Element.isElement });
   expect(node).toEqual(output);
 });