+++ packages/slate-plugins/src/__tests__/deserializers/deserialize-html/deserializeElement/plugin-p__element-div__type-p.spec.tsx
@@ -14,9 +14,9 @@ const input = {
 };
 
 const output = (
-  <p>
-    <text>test</text>
-  </p>
+  <hp>
+    <htext>test</htext>
+  </hp>
 );
 
 it('should be', () => {