+++ packages/slate-plugins/src/__tests__/deserializers/deserialize-html/deserializeFragment/body.spec.tsx
@@ -5,9 +5,9 @@ import { deserializeFragment } from 'deserializers/deserialize-html/utils';
 
 const fragment = (
   <fragment>
-    <p>
+    <hp>
       <htext>test</htext>
-    </p>
+    </hp>
   </fragment>
 );