+++ packages/slate-plugins/src/__tests__/element/withBreakEmptyReset/text-empty-onUnwrap.spec.tsx
@@ -7,10 +7,10 @@ import { Editor } from 'slate';
 
 const input = (
   <editor>
-    <blockquote>
-      <text />
+    <hblockquote>
+      <htext />
       <cursor />
-    </blockquote>
+    </hblockquote>
   </editor>
 ) as any;