+++ packages/slate-plugins/src/__tests__/element/withDeleteStartReset/collapsed.spec.tsx
@@ -7,18 +7,18 @@ import { Editor } from 'slate';
 
 const input = (
   <editor>
-    <blockquote>
-      <text />
+    <hblockquote>
+      <htext />
       <cursor />
-    </blockquote>
+    </hblockquote>
   </editor>
 ) as any;
 
 const output = (
   <editor>
-    <p>
-      <text />
-    </p>
+    <hp>
+      <htext />
+    </hp>
   </editor>
 ) as any;