+++ packages/slate-plugins/src/__tests__/elements/link/withLink/text-url.spec.tsx
@@ -6,10 +6,10 @@ import { withReact } from 'slate-react';
 
 const input = (
   <editor>
-    <p>
+    <hp>
       test
       <cursor />
-    </p>
+    </hp>
   </editor>
 ) as any;
 
@@ -17,13 +17,13 @@ const text = 'http://localhost:3000';
 
 const output = (
   <editor>
-    <p>
+    <hp>
       test
       <element type="a" url="http://localhost:3000">
         http://localhost:3000
       </element>
-      <text />
-    </p>
+      <htext />
+    </hp>
   </editor>
 ) as any;