+++ packages/slate-plugins/src/__tests__/soft-break/onKeyDownSoftBreak/valid.spec.tsx
@@ -6,10 +6,10 @@ import { onKeyDownSoftBreak } from 'soft-break';
 
 const input = (
   <editor>
-    <p>
+    <hp>
       test
       <cursor />
-    </p>
+    </hp>
   </editor>
 ) as any;
 
@@ -17,10 +17,10 @@ const event = new KeyboardEvent('keydown');
 
 const output = (
   <editor>
-    <p>
+    <hp>
       test{'\n'}
       <cursor />
-    </p>
+    </hp>
   </editor>
 ) as any;